[Feature #153114603]Rework template
Previously the main object rendered in the template was form object
for creating post categories. The issue was that when there were errors
in post form data submitted from the same template those data would be lost
as they were sent to another view.

This commit will change how it works. The main object will be form for creating
posts. If a user made any error when submitting data, the form will show the errors
and the data inputs in each field will not be lost.